requirements.txt
openai-whisper
spacy==3.7.4
sentence-transformers
faiss-cpu
streamlit
jinja2
pandas
numpy
fastapi
uvicorn[standard]
python-multipart      # required for file uploads
pydantic



